refactor: 약관 및 정책 이동 버튼 client 컴포넌트로 분리 (#21)
