+++ apps/web/src/app/mypage/policies/PolicyButtonClient/PolicyButtonClient.style.ts
@@ -1,24 +1,5 @@
 import styled from '@emotion/styled';
 
-export const Wrapper = styled.main`
-  display: flex;
-  flex-direction: column;
-  min-height: 100dvh;
-  background-color: #242426;
-`;
-
-export const SrOnly = styled.h1`
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
-  border: 0;
-`;
-
 export const ContentLayout = styled.div`
   padding: 16px;
 `;
